+++ lib/types/createPaymentMethodOpts.dart
@@ -1,9 +1,7 @@
 import 'package:appstitch_stripe/types/address.dart';
-import 'package:appstitch_stripe/types/cardFormPaymentRequest.dart';
 import 'package:appstitch_stripe/types/connectOptions.dart';
 import 'package:appstitch_stripe/types/createCardOpts.dart';
 import 'package:json_annotation/json_annotation.dart';
-import 'package:appstitch_stripe/types/token.dart';
 
 part 'createPaymentMethodOpts.g.dart';