+++ lib/types/createPaymentMethodOpts.g.dart
@@ -16,33 +16,25 @@ CreatePaymentMethodOpts _$CreatePaymentMethodOptsFromJson(
         ? null
         : CreateCardOpts.fromJson(json['card'] as Map<String, dynamic>),
   )
-    ..blueprintId = json['blueprint_id'] as String
+    ..blueprintId = json['blueprintId'] as String
     ..id = json['id'] as String
     ..stripeError = json['type'] as String
     ..message = json['message'] as String
     ..param = json['param'] as String
     ..object = json['object'] as String
-    ..stripeAccount = json['stripe_account'] as String;
+    ..stripeAccount = json['stripeAccount'] as String;
 }
 
 Map<String, dynamic> _$CreatePaymentMethodOptsToJson(
-    CreatePaymentMethodOpts instance) {
-  final val = <String, dynamic>{};
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull('blueprint_id', instance.blueprintId);
-  writeNotNull('id', instance.id);
-  writeNotNull('type', instance.stripeError);
-  writeNotNull('message', instance.message);
-  writeNotNull('param', instance.param);
-  writeNotNull('object', instance.object);
-  writeNotNull('stripe_account', instance.stripeAccount);
-  val['billingAddress'] = instance.billingAddress;
-  val['card'] = instance.card;
-  return val;
-}
+        CreatePaymentMethodOpts instance) =>
+    <String, dynamic>{
+      'blueprintId': instance.blueprintId,
+      'id': instance.id,
+      'type': instance.stripeError,
+      'message': instance.message,
+      'param': instance.param,
+      'object': instance.object,
+      'stripeAccount': instance.stripeAccount,
+      'billingAddress': instance.billingAddress,
+      'card': instance.card,
+    };