+++ lib/types/token.dart
@@ -1,20 +1,24 @@
+import 'package:appstitch_stripe/types.dart';
 import 'package:json_annotation/json_annotation.dart';
-import 'bankAccount.dart';
-import 'card.dart';
 part 'token.g.dart';
 
 @JsonSerializable(
     explicitToJson: true, includeIfNull: false, fieldRename: FieldRename.snake)
 class Token {
-  BankAccount bankAccount;
-  CreditCard card;
+  CreateCardOpts card;
   double created;
   bool livemode;
   String tokenId;
 
-  Token(
-      {this.bankAccount, this.card, this.created, this.livemode, this.tokenId});
+  Token({this.card, this.created, this.livemode, this.tokenId});
 
-  factory Token.fromJson(Map<String, dynamic> json) => _$TokenFromJson(json);
+  factory Token.fromJson(Map<dynamic, dynamic> json) {
+    return Token(
+        card:
+            json['card'] != null ? CreateCardOpts.fromJson(json['card']) : null,
+        created: json['created'],
+        tokenId: json['tokenId'],
+        livemode: json["livemode"]);
+  }
   Map<String, dynamic> toJson() => _$TokenToJson(this);
 }